Move FreeTrialHandler.js to new VaultV2 directory
