+++ modules/ppcp-button/resources/js/VaultV2/FreeTrialHandler.js
